+++ packages/@headlessui-react/src/components/focus-trap/focus-trap.tsx
@@ -23,8 +23,8 @@ import { useEventListener } from '../../hooks/use-event-listener'
 import { microTask } from '../../utils/micro-task'
 import { useWatch } from '../../hooks/use-watch'
 import { useDisposables } from '../../hooks/use-disposables'
-import { onDocumentReady } from '../../utils/document-ready'
 import { useOnUnmount } from '../../hooks/use-on-unmount'
+import { history } from '../../utils/active-element-history'
 
 type Containers =
   // Lazy resolved containers
@@ -211,29 +211,6 @@ export let FocusTrap = Object.assign(FocusTrapRoot, {
 })
 
 // ---
-
-let history: HTMLElement[] = []
-onDocumentReady(() => {
-  function handle(e: Event) {
-    if (!(e.target instanceof HTMLElement)) return
-    if (e.target === document.body) return
-    if (history[0] === e.target) return
-
-    history.unshift(e.target)
-
-    // Filter out DOM Nodes that don't exist anymore
-    history = history.filter((x) => x != null && x.isConnected)
-    history.splice(10) // Only keep the 10 most recent items
-  }
-
-  window.addEventListener('click', handle, { capture: true })
-  window.addEventListener('mousedown', handle, { capture: true })
-  window.addEventListener('focus', handle, { capture: true })
-
-  document.body.addEventListener('click', handle, { capture: true })
-  document.body.addEventListener('mousedown', handle, { capture: true })
-  document.body.addEventListener('focus', handle, { capture: true })
-})
 
 function useRestoreElement(enabled: boolean = true) {
   let localHistory = useRef<HTMLElement[]>(history.slice())
