+++ packages/@headlessui-vue/src/components/focus-trap/focus-trap.ts
@@ -22,7 +22,7 @@ import { useTabDirection, Direction as TabDirection } from '../../hooks/use-tab-
 import { getOwnerDocument } from '../../utils/owner'
 import { useEventListener } from '../../hooks/use-event-listener'
 import { microTask } from '../../utils/micro-task'
-import { onDocumentReady } from '../../utils/document-ready'
+import { history } from '../../utils/active-element-history'
 
 type Containers =
   // Lazy resolved containers
@@ -208,29 +208,6 @@ export let FocusTrap = Object.assign(
   }),
   { features: Features }
 )
-
-let history: HTMLElement[] = []
-onDocumentReady(() => {
-  function handle(e: Event) {
-    if (!(e.target instanceof HTMLElement)) return
-    if (e.target === document.body) return
-    if (history[0] === e.target) return
-
-    history.unshift(e.target)
-
-    // Filter out DOM Nodes that don't exist anymore
-    history = history.filter((x) => x != null && x.isConnected)
-    history.splice(10) // Only keep the 10 most recent items
-  }
-
-  window.addEventListener('click', handle, { capture: true })
-  window.addEventListener('mousedown', handle, { capture: true })
-  window.addEventListener('focus', handle, { capture: true })
-
-  document.body.addEventListener('click', handle, { capture: true })
-  document.body.addEventListener('mousedown', handle, { capture: true })
-  document.body.addEventListener('focus', handle, { capture: true })
-})
 
 function useRestoreElement(enabled: Ref<boolean>) {
   let localHistory = ref<HTMLElement[]>(history.slice())
